Created DSLGenerator and CodeGenerator
Those are xtend files needed to generate the file structure and the
various api files we have, together with the primary and secondary
control scripts for the bricks. Essentially all the code we need
for the rover.
So far only the API files are generated properly since we need no
dynamic changes for them as they don't directly interact with the
variables changed using the defined grammar.