WIP: All API files are generated + secondary and some of primary
It seems that all the files are generated properly yet they have
not been tested on the rover itself.

The main file still misses most of the functionality. Only the
imports, global variables and conditional generation of the
report file code is done. The main and other parts are missing.