WIP: Added code to CodeGenerator.xtend but not modularized
All the code was added to CodeGenerator.xtend file but it's not
fully modularized.
More precisely, generateMain(), generateMissionOngoing() and all
the GenerateFunctions(*) methods still need to be modified so that
they only add the functionality that is needed. So all these
methods will just generate all the code that should be in
AdvancedRover.py if all the functionalty is present but so far it
does not distinguish between what should and shouldn't be added
based on the specification file (spec1.mr). This functionality
still needs to be added.